+++ dataland-backend-utils/src/main/kotlin/org/dataland/datalandbackendutils/exceptions/ResourceNotFoundApiException.kt
@@ -7,7 +7,7 @@ import org.springframework.http.HttpStatus
  * A ResourceNotFoundApiException should be thrown if a dataland-internal resource could not be located
  * Both message and summary are displayed to the user with a 404 status code
  */
-class ResourceNotFoundApiException(
+open class ResourceNotFoundApiException(
     val summary: String,
     override val message: String,
     cause: Throwable? = null,
